Changed Float to Double in definition of Table element. (Double is more efficient in GHC.) Truncate width in opendocument output to 2 decimal places.
git-svn-id: https://pandoc.googlecode.com/svn/trunk@1418 788f1e2b-df1e-0410-8736-df70ead52e1b

+++ Text/Pandoc/Definition.hs
@@ -84,7 +84,7 @@ data Block
                             -- the term, and a block list)
     | Header Int [Inline]   -- ^ Header - level (integer) and text (inlines) 
     | HorizontalRule        -- ^ Horizontal rule
-    | Table [Inline] [Alignment] [Float] [[Block]] [[[Block]]]  -- ^ Table,
+    | Table [Inline] [Alignment] [Double] [[Block]] [[[Block]]]  -- ^ Table,
                             -- with caption, column alignments,
                             -- relative column widths, column headers
                             -- (each a list of blocks), and rows
